refactor(modules): Move evaluation modules into core
