+++ .github/release-please/.release-please-manifest.json
@@ -3,6 +3,5 @@
     "cmd/wasm": "0.1.2",
     "openfeature/providers/python-provider": "0.5.0",
     "openfeature/providers/kotlin-provider": "0.3.0",
-    "modules/evaluation": "0.2.2",
     "modules/core": "0.4.0"
 }
