--- modules/evaluation/CHANGELOG.md
@@ -1,61 +1,0 @@
-# Changelog
-
-## [0.2.2](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.2.1...modules/evaluation/v0.2.2) (2026-01-19)
-
-
-### 🔧 Chores
-
-* Bump github.com/aws/aws-sdk-go-v2/service/s3 ([#4613](https://github.com/thomaspoignant/go-feature-flag/issues/4613)) ([a8e9b13](https://github.com/thomaspoignant/go-feature-flag/commit/a8e9b136eb972da636726525bffc1c2f86fcb432))
-* **dependency:** Bump module modules/core 0.4.0 ([#4659](https://github.com/thomaspoignant/go-feature-flag/issues/4659)) ([2140385](https://github.com/thomaspoignant/go-feature-flag/commit/21403850eeb13b9b5cd7e85b226f6a10874352ab))
-
-## [0.2.1](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.2.0...modules/evaluation/v0.2.1) (2026-01-09)
-
-
-### 🔧 Chores
-
-* Bump k8s.io/api from 0.34.2 to 0.34.3 ([#4464](https://github.com/thomaspoignant/go-feature-flag/issues/4464)) ([7c947e6](https://github.com/thomaspoignant/go-feature-flag/commit/7c947e68ec359dbf2afbb1ac80ccbd31c56982cb))
-* Bump k8s.io/api from 0.34.3 to 0.35.0 ([#4517](https://github.com/thomaspoignant/go-feature-flag/issues/4517)) ([631144b](https://github.com/thomaspoignant/go-feature-flag/commit/631144b01a33c2531ab4d2160be908d96a80347f))
-* **dependency:** Bump module modules/core 0.3.1 ([#4586](https://github.com/thomaspoignant/go-feature-flag/issues/4586)) ([f79d92e](https://github.com/thomaspoignant/go-feature-flag/commit/f79d92e36642ff67fdbd15eb60568c3e7d24814a))
-
-## [0.2.0](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.1.3...modules/evaluation/v0.2.0) (2025-12-09)
-
-
-### 🚀 New Features
-
-* Support x-api-key header for authentication ([#4347](https://github.com/thomaspoignant/go-feature-flag/issues/4347)) ([3ca07a8](https://github.com/thomaspoignant/go-feature-flag/commit/3ca07a8fa49522aa8b348bb5314a6f503dfa9778))
-
-
-### 🔧 Chores
-
-* Bump github.com/aws/aws-sdk-go-v2/service/s3 ([#4263](https://github.com/thomaspoignant/go-feature-flag/issues/4263)) ([3944a49](https://github.com/thomaspoignant/go-feature-flag/commit/3944a491413056d903236573fbf5a75fc7336dd9))
-* Bump github.com/aws/aws-sdk-go-v2/service/s3 ([#4326](https://github.com/thomaspoignant/go-feature-flag/issues/4326)) ([927e392](https://github.com/thomaspoignant/go-feature-flag/commit/927e392662eaad75f33bd88275c566a465d02446))
-* **dependency:** Bump module modules/core 0.3.0 ([#4441](https://github.com/thomaspoignant/go-feature-flag/issues/4441)) ([a1f00c5](https://github.com/thomaspoignant/go-feature-flag/commit/a1f00c5619cbb9741617ec0ba2a760e02e929451))
-
-## [0.1.3](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.1.2...modules/evaluation/v0.1.3) (2025-11-04)
-
-
-### 🔧 Chores
-
-* **dependencies:** Bump github.com/thomaspoignant/go-feature-flag/modules/core ([#4247](https://github.com/thomaspoignant/go-feature-flag/issues/4247)) ([ab482c2](https://github.com/thomaspoignant/go-feature-flag/commit/ab482c224cfcf6298b6f0fc6e96fd53f6f2171e1))
-
-## [0.1.2](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.1.1...modules/evaluation/v0.1.2) (2025-10-14)
-
-
-### 🔧 Chores
-
-* **dependency:** Bump module  ${MODULE_PATH} ${VERSION} ([#4094](https://github.com/thomaspoignant/go-feature-flag/issues/4094)) ([36c2fe9](https://github.com/thomaspoignant/go-feature-flag/commit/36c2fe9dc19b22e961b97595b2f033ebd3cc2883))
-* **repo:** Create an evaluation go module ([#4079](https://github.com/thomaspoignant/go-feature-flag/issues/4079)) ([2305959](https://github.com/thomaspoignant/go-feature-flag/commit/230595939b35e9472a422e0b265fb450b20d3651))
-
-## [0.1.1](https://github.com/thomaspoignant/go-feature-flag/compare/modules/evaluation/v0.1.0...modules/evaluation/v0.1.1) (2025-10-13)
-
-
-### 🐛 Bug Fixes
-
-* test triggering CI for releasing GO module ([#4063](https://github.com/thomaspoignant/go-feature-flag/issues/4063)) ([bae6b71](https://github.com/thomaspoignant/go-feature-flag/commit/bae6b719f1d5ce4ec64a1498ee2630255bbdf49c))
-
-## 0.1.0 (2025-10-10)
-
-
-### 🔧 Chores
-
-* initial release please setup ([#4028](https://github.com/thomaspoignant/go-feature-flag/issues/4028)) ([081e1ab](https://github.com/thomaspoignant/go-feature-flag/commit/081e1aba45f7d32073802ddceb3790766c6ef4ea))
