--- modules/evaluation/go.mod
@@ -1,23 +1,0 @@
-module github.com/thomaspoignant/go-feature-flag/modules/evaluation
-
-go 1.24.6
-
-require (
-	github.com/stretchr/testify v1.11.1
-	github.com/thomaspoignant/go-feature-flag/modules/core v0.4.0
-)
-
-require (
-	github.com/GeorgeD19/json-logic-go v0.0.0-20220225111652-48cc2d2c387e // indirect
-	github.com/antlr4-go/antlr/v4 v4.13.0 // indirect
-	github.com/blang/semver v3.5.1+incompatible // indirect
-	github.com/buger/jsonparser v1.1.1 // indirect
-	github.com/dariubs/percent v0.0.0-20190521174708-8153fcbd48ae // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/nikunjy/rules v1.5.0 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	github.com/spf13/cast v1.3.0 // indirect
-	golang.org/x/exp v0.0.0-20240506185415-9bf2ced13842 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
